requirements.txt
pymysql
requests
django-floppyforms